refactoring notary page and tables
