+++ src/components/OverviewApp/Notary/index.ts
@@ -1,0 +1,7 @@
+import CancelProposalTable from "./CancelProposalTable"
+import LargeRequestTable from "./LargeRequestTable"
+import NotaryTabs from "./NotaryTabs"
+import PublicRequestTable from "./PublicRequestTable"
+import VerifiedClientsTable from "./VerifiedClientsTable"
+
+export { CancelProposalTable, LargeRequestTable, PublicRequestTable, VerifiedClientsTable, NotaryTabs }
